requirements.txt
django>=1.11,<1.11.99
mysqlclient>=1.3,<1.3.99
gunicorn>=19.7,<19.7.99
factory_boy>=2.8,<2.8.99
django-formtools>=2.0,<2.0.99
django-filter>=1.0,<1.0.99
django-annoying>=0.10,<0.10.99
reportlab>=3.4,<3.4.99
django-leaflet>=0.22,<0.22.99
django-extra-views>=0.9,<0.9.99
transifex-client>=0.12,<0.12.99
django-template-i18n-lint>=1.2,<1.2.99
pylabels>=1.2,<1.2.99
django-avatar>=3.1,<3.1.99
django-localflavor>=1.4,<1.4.99
rules>=1.2,<1.2.99
